+++ src/main/java/com/contentful/java/cda/CDAClient.java
@@ -1,12 +1,20 @@
 package com.contentful.java.cda;
 
+import com.contentful.java.cda.interceptor.ErrorInterceptor;
+import com.contentful.java.cda.interceptor.HeaderInterceptor;
+import com.contentful.java.cda.interceptor.LogInterceptor;
+
+import java.util.HashMap;
 import java.util.Map;
+import java.util.Properties;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.Executor;
-import retrofit.RestAdapter;
-import retrofit.RestAdapter.LogLevel;
-import retrofit.client.Client;
-import retrofit.client.Response;
+
+import okhttp3.OkHttpClient;
+import retrofit2.Response;
+import retrofit2.Retrofit;
+import retrofit2.adapter.rxjava.RxJavaCallAdapterFactory;
+import retrofit2.converter.gson.GsonConverterFactory;
 import rx.Observable;
 import rx.functions.Func1;
 
@@ -54,39 +62,56 @@ public class CDAClient {
       endpoint = ENDPOINT_PROD;
     }
 
-    RestAdapter.Builder restBuilder = new RestAdapter.Builder()
-        .setEndpoint(endpoint)
-        .setRequestInterceptor(new Interceptor(token));
+    OkHttpClient.Builder okBuilder = new OkHttpClient.Builder()
+        .addInterceptor(new HeaderInterceptor(createUserAgent(), token))
+        .addInterceptor(new ErrorInterceptor());
+    okBuilder = setLogger(okBuilder, clientBuilder);
 
-    setLogLevel(restBuilder, clientBuilder);
-    setLog(restBuilder, clientBuilder);
-    setClient(restBuilder, clientBuilder);
-    return restBuilder.build().create(CDAService.class);
+    Retrofit.Builder retrofitBuilder = new Retrofit.Builder()
+        .addConverterFactory(GsonConverterFactory.create(ResourceFactory.GSON))
+        .addCallAdapterFactory(RxJavaCallAdapterFactory.create())
+        .baseUrl(endpoint);
+
+    clientBuilder.setClient(okBuilder.build());
+
+    retrofitBuilder = setClient(retrofitBuilder, clientBuilder);
+
+    return retrofitBuilder.build().create(CDAService.class);
   }
 
-  private void setClient(RestAdapter.Builder restBuilder, Builder clientBuilder) {
+  private Retrofit.Builder setClient(Retrofit.Builder retrofitBuilder, Builder clientBuilder) {
     if (clientBuilder.client != null) {
-      restBuilder.setClient(clientBuilder.client);
+      return retrofitBuilder.client(clientBuilder.client);
     }
+    return retrofitBuilder;
   }
 
-  private void setLogLevel(RestAdapter.Builder restBuilder, Builder clientBuilder) {
-    if (clientBuilder.logLevel != null) {
-      restBuilder.setLogLevel(clientBuilder.logLevel);
+  private OkHttpClient.Builder setLogger(OkHttpClient.Builder okBuilder, Builder clientBuilder) {
+    if (clientBuilder.logger != null) {
+      switch (clientBuilder.logLevel) {
+        case NONE:
+          break;
+        case BASIC:
+          okBuilder.addInterceptor(new LogInterceptor(clientBuilder.logger));
+          break;
+        case FULL:
+          okBuilder.addNetworkInterceptor(new LogInterceptor(clientBuilder.logger));
+          break;
+      }
+    } else {
+      if (clientBuilder.logLevel != Logger.Level.NONE) {
+        throw new IllegalArgumentException("Cannot log to a null logger. Please set either logLevel to None, or do set a Logger");
+      }
     }
-  }
-
-  private void setLog(RestAdapter.Builder restBuilder, Builder clientBuilder) {
-    if (clientBuilder.log != null) {
-      restBuilder.setLog(clientBuilder.log);
-    }
+    return okBuilder;
   }
 
   /**
    * Returns a {@link FetchQuery} for a given {@code type}, which can be used to fulfill the
    * request synchronously or asynchronously when a callback is provided.
+   *
    * @param type resource type.
-   * @param <T> resource type.
+   * @param <T>  resource type.
    * @return query instance.
    */
   public <T extends CDAResource> FetchQuery<T> fetch(Class<T> type) {
@@ -96,8 +121,9 @@ public class CDAClient {
   /**
    * Returns an {@link ObserveQuery} for a given {@code type}, which can be used to return
    * an {@link Observable} that fetches the desired resources.
+   *
    * @param type resource type.
-   * @param <T> resource type.
+   * @param <T>  resource type.
    * @return query instance.
    */
   public <T extends CDAResource> ObserveQuery<T> observe(Class<T> type) {
@@ -106,6 +132,7 @@ public class CDAClient {
 
   /**
    * Returns a {@link SyncQuery} for initial synchronization via the Sync API.
+   *
    * @return query instance.
    * @throws UnsupportedOperationException if tried to sync with a preview token
    */
@@ -116,6 +143,7 @@ public class CDAClient {
   /**
    * Returns a {@link SyncQuery} for synchronization with the provided {@code syncToken} via
    * the Sync API.
+   *
    * @param syncToken sync token.
    * @return query instance.
    * @throws UnsupportedOperationException if tried to sync with a preview token
@@ -126,6 +154,7 @@ public class CDAClient {
 
   /**
    * Returns a {@link SyncQuery} for synchronization with an existing space.
+   *
    * @param synchronizedSpace space to sync.
    * @return query instance.
    * @throws UnsupportedOperationException if tried to sync with a preview token
@@ -149,23 +178,31 @@ public class CDAClient {
     return builder.build();
   }
 
-  /** Fetches the space for this client (synchronously). */
+  /**
+   * Fetches the space for this client (synchronously).
+   */
   public CDASpace fetchSpace() {
     return observeSpace().toBlocking().first();
   }
 
-  /** Fetches the space for this client (asynchronously). */
+  /**
+   * Fetches the space for this client (asynchronously).
+   */
   @SuppressWarnings("unchecked")
   public <C extends CDACallback<CDASpace>> C fetchSpace(C callback) {
     return (C) Callbacks.subscribeAsync(observeSpace(), callback, this);
   }
 
-  /** Returns an {@link Observable} that fetches the space for this client. */
+  /**
+   * Returns an {@link Observable} that fetches the space for this client.
+   */
   public Observable<CDASpace> observeSpace() {
     return cacheSpace(true);
   }
 
-  /** Caching */
+  /**
+   * Caching
+   */
   Observable<Cache> cacheAll(final boolean invalidate) {
     return cacheSpace(invalidate)
         .flatMap(new Func1<CDASpace, Observable<Map<String, CDAContentType>>>() {
@@ -183,11 +220,11 @@ public class CDAClient {
   Observable<CDASpace> cacheSpace(boolean invalidate) {
     CDASpace space = invalidate ? null : cache.space();
     if (space == null) {
-      return service.space(spaceId).map(new Func1<Response, CDASpace>() {
-        @Override public CDASpace call(Response response) {
-          CDASpace tmp = ResourceFactory.space(response);
-          cache.setSpace(tmp);
-          return tmp;
+      return service.space(spaceId).map(new Func1<Response<CDASpace>, CDASpace>() {
+        @Override public CDASpace call(Response<CDASpace> response) {
+          CDASpace space = ResourceFactory.space(response);
+          cache.setSpace(space);
+          return space;
         }
       });
     }
@@ -197,10 +234,10 @@ public class CDAClient {
   Observable<Map<String, CDAContentType>> cacheTypes(boolean invalidate) {
     Map<String, CDAContentType> types = invalidate ? null : cache.types();
     if (types == null) {
-      return service.array(spaceId, PATH_CONTENT_TYPES, null).map(
-          new Func1<Response, Map<String, CDAContentType>>() {
-            @Override public Map<String, CDAContentType> call(Response response) {
-              CDAArray array = ResourceFactory.array(response, CDAClient.this);
+      return service.array(spaceId, PATH_CONTENT_TYPES, new HashMap<String, String>()).map(
+          new Func1<Response<CDAArray>, Map<String, CDAContentType>>() {
+            @Override public Map<String, CDAContentType> call(Response<CDAArray> arrayResponse) {
+              CDAArray array = ResourceFactory.array(arrayResponse, CDAClient.this);
               Map<String, CDAContentType> tmp = new ConcurrentHashMap<String, CDAContentType>();
               for (CDAResource resource : array.items()) {
                 tmp.put(resource.id(), (CDAContentType) resource);
@@ -228,7 +265,20 @@ public class CDAClient {
     return Observable.just(contentType);
   }
 
-  /** Returns a {@link CDAClient} builder. */
+  String createUserAgent() {
+    final Properties properties = System.getProperties();
+    return String.format("contentful.java/%s(%s %s) %s/%s",
+        Util.getProperty("version.name"),
+        properties.getProperty("java.runtime.name"),
+        properties.getProperty("java.runtime.version"),
+        properties.getProperty("os.name"),
+        properties.getProperty("os.version")
+    );
+  }
+
+  /**
+   * Returns a {@link CDAClient} builder.
+   */
   public static Builder builder() {
     return new Builder();
   }
@@ -243,55 +293,74 @@ public class CDAClient {
     String space;
     String token;
     String endpoint;
-    LogLevel logLevel;
-    RestAdapter.Log log;
-    Client client;
+
+    Logger logger;
+    Logger.Level logLevel = Logger.Level.NONE;
+
+    OkHttpClient client;
     boolean preview;
 
-    /** Sets the space ID. */
+    /**
+     * Sets the space ID.
+     */
     public Builder setSpace(String space) {
       this.space = space;
       return this;
     }
 
-    /** Sets the space access token. */
+    /**
+     * Sets the space access token.
+     */
     public Builder setToken(String token) {
       this.token = token;
       return this;
     }
 
-    /** Sets a custom endpoint. */
+    /**
+     * Sets a custom endpoint.
+     */
     public Builder setEndpoint(String endpoint) {
       this.endpoint = endpoint;
       return this;
     }
 
-    /** Sets a custom log level. */
-    public Builder setLogLevel(LogLevel logLevel) {
+    /**
+     * Sets a custom logger level.
+     * <p>
+     * If set to {@link Logger.Level}.NONE any custom logger will get ignored.
+     */
+    public Builder setLogLevel(Logger.Level logLevel) {
       this.logLevel = logLevel;
       return this;
     }
 
-    /** Sets a custom logger. */
-    public Builder setLog(RestAdapter.Log log) {
-      this.log = log;
+    /**
+     * Sets a custom logger.
+     */
+    public Builder setLogger(Logger logger) {
+      this.logger = logger;
       return this;
     }
 
-    /** Sets the endpoint to point the Preview API. */
+    /**
+     * Sets the endpoint to point the Preview API.
+     */
     public Builder preview() {
       preview = true;
       return this.setEndpoint(Constants.ENDPOINT_PREVIEW);
     }
 
-    /** Sets a custom HTTP client. */
-    public Builder setClient(Client client) {
+    /**
+     * Sets a custom HTTP client.
+     */
+    public Builder setClient(OkHttpClient client) {
       this.client = client;
       return this;
     }
 
     /**
      * Create CDAClient, using the specified configuration options.
+     *
      * @return a build CDAClient.
      */
     public CDAClient build() {