+++ src/main/java/com/contentful/java/cda/Platform.java
@@ -1,8 +1,10 @@
 package com.contentful.java.cda;
 
 import android.os.Build;
+import android.os.Handler;
+import android.os.Looper;
+
 import java.util.concurrent.Executor;
-import retrofit.android.MainThreadExecutor;
 
 abstract class Platform {
   private static final Platform PLATFORM = findPlatform();
@@ -25,17 +27,27 @@ abstract class Platform {
 
   abstract Executor callbackExecutor();
 
-  /** Provides sane defaults for operation on the JVM. */
+  /**
+   * Provides sane defaults for operation on the JVM.
+   */
   private static class Base extends Platform {
     @Override Executor callbackExecutor() {
       return new SynchronousExecutor();
     }
   }
 
-  /** Provides sane defaults for operation on Android. */
+  /**
+   * Provides sane defaults for operation on Android.
+   */
   private static class Android extends Platform {
     @Override Executor callbackExecutor() {
-      return new MainThreadExecutor();
+      return new Executor() {
+        private final Handler handler = new Handler(Looper.getMainLooper());
+
+        @Override public void execute(Runnable command) {
+          handler.post(command);
+        }
+      };
     }
   }
 }